[FIX] util/models: Fix missing column binding_model_id
Issue summary
This issue is caused by a migration 3e22c9a that removes automated actions
before removing a model.
However, the migration assumes the presence of the binding_model_id field on
ir_act_server, which is not available in older versions.

Root cause (version mismatch)
- "base.action.rule" renamed to "base.automation" in: 10.saas~14.1.0 0eea7ea
- action_server_id field on base_automation
Introduced in: 10.saas~14.1.3 0eea7ea
- binding_model_id field on ir_act_server
Introduced later in: 10.saas~18.1.3 609e381
- The customer database is on 10.saas~14.1.3, where:
- base_automation.action_server_id exists
- But ir_act_server.binding_model_id does NOT exist as it introduced in
10.saas~18.1.3

During migration, when upgrading past 10.saas~17.1.3, the script:
```/home/labo/src/upgrade/migrations/base/10.saas~17.1.3/pre-20-report.py```

Calls :- util.remove_model(cr, report) And the issue arise of field not exist.
Soln :- Add an additional column existence check before executing the DELETE:

```
2026-01-29 08:58:43,708 19 INFO labo_3858705_19.0 odoo.modules.migration: module base: Running migration [>10.saas~17.1.3] pre-20-report
2026-01-29 08:58:44,288 19 ERROR labo_3858705_19.0 odoo.sql_db: bad query:
                DELETE FROM base_automation b_a
                 USING ir_act_server action
                 WHERE b_a.action_server_id = action.id
                   AND action.binding_model_id = 154

ERROR: column action.binding_model_id does not exist
LINE 5:                    AND action.binding_model_id = 154
                               ^

2026-01-29 08:58:44,290 19 WARNING labo_3858705_19.0 odoo.modules.loading: Transient module states were reset
2026-01-29 08:58:44,290 19 ERROR labo_3858705_19.0 odoo.modules.registry: Failed to load registry
Traceback (most recent call last):
  File /home/odoo/src/odoo/11.0/odoo/modules/registry.py, line 85, in new
    odoo.modules.load_modules(registry._db, force_demo, status, update_module)
  File /home/odoo/src/odoo/11.0/odoo/modules/loading.py, line 324, in load_modules
    report=report, models_to_check=models_to_check)
  File /home/odoo/src/odoo/11.0/odoo/modules/loading.py, line 135, in load_module_graph
    migrations.migrate_module(package, 'pre')
  File /home/odoo/src/odoo/11.0/odoo/modules/migration.py, line 202, in migrate_module
    migrate(self.cr, installed_version)
  File /home/odoo/src/odoo/11.0/odoo/addons/base/maintenance/migrations/base/10.saas~17.1.3/pre-20-report.py, line 16, in migrate
    util.remove_model(cr, report)
  File /home/odoo/src/odoo/11.0/odoo/addons/base/maintenance/migrations/util/models.py, line 188, in remove_model
    [mod_id],
  File /home/odoo/src/odoo/11.0/odoo/sql_db.py, line 155, in wrapper
    return f(self, *args, **kwargs)
  File /home/odoo/src/odoo/11.0/odoo/sql_db.py, line 232, in execute
    res = self._obj.execute(query, params)
psycopg2.errors.UndefinedColumn: column action.binding_model_id does not exist
LINE 5:                    AND action.binding_model_id = 154
```
UPG:- 3858705
OPW:- 5883894

+++ src/util/models.py
@@ -177,7 +177,11 @@ def remove_model(cr, model, drop_table=True, ignore_m2m=()):
             if column_exists(cr, tbl, "model_id"):
                 cr.execute("DELETE FROM {0} WHERE model_id=%s".format(tbl), [mod_id])
 
-        if not version_gte("17.0") and column_exists(cr, "base_automation", "action_server_id"):
+        if (
+            not version_gte("17.0")
+            and column_exists(cr, "base_automation", "action_server_id")
+            and column_exists(cr, "ir_act_server", "binding_model_id")
+        ):
             cr.execute(
                 """
                 DELETE FROM base_automation b_a
